Rename overkiz_entity.py
